requirements.txt
anyio==3.6.1
certifi==2022.6.15
charset-normalizer==2.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
python-dotenv==0.20.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.2.0
urllib3==1.26.11
